+++ backend/src/users/users.service.ts
@@ -4,15 +4,36 @@ import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Profile } from '../profiles/entities/profiles.entity';
 
 @Injectable()
 export class UsersService {
   constructor(
-    @InjectRepository(User) private userRepository: Repository<User>,
+    @InjectRepository(User)
+    private readonly userRepository: Repository<User>,
+
+    @InjectRepository(Profile)
+    private readonly profileRepository: Repository<Profile>,
   ) {}
 
   async create(createUserDto: CreateUserDto): Promise<User> {
-    const user = this.userRepository.create(createUserDto);
+    const profile = this.profileRepository.create({
+      description: 'New user profile',
+      followers: 0,
+      following: 0,
+      likes: 0,
+      visits: 0,
+      comments: 0,
+      posts_number: 0,
+    });
+
+    const savedProfile = await this.profileRepository.save(profile);
+
+    const user = this.userRepository.create({
+      ...createUserDto,
+      profile: savedProfile,
+    });
+
     return await this.userRepository.save(user);
   }
 
@@ -35,7 +56,15 @@ export class UsersService {
   }
 
   async remove(id: number): Promise<User> {
-    const user = await this.findOne(id);
+    const user = await this.userRepository.findOne({
+      where: { id },
+      relations: {
+        profile: true,
+      },
+    });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
     return await this.userRepository.remove(user);
   }
 }
